Define evil-leader register bindings
Define pseudo evil-leader bindings for "+ and "* so that specifying the
clipboard or current selection is easier to type. This requires (it
seems) simulating keypresses because these sequences are usually part of
a larger evil sequence, or maybe I just haven't found the simple way to
do it.

An alternative is to define some bindings for different combinations of
register, commands, and motions, but a composable solution is
preferable.