[i3] Move Super keys to mod3
For some reason both Super and Hyper keys are mapped to mod4.
This modmap removes Super keys from mod4 and adds them to mod3.
This allows use of super keys separate from hyper keys.
Since i3 is using Hyper for its modifier, this allows the super
keys to be used for other purposes.