[i3] Move Super_L, Super_R to mod3
Since I am using Hyper for i3 key, I want to use Super_L and Super_R
for other purposes. The default mapping is mod4 has Hyper's and Super's,
so it seems reasonable to move Super to mod3 to achieve this. Xmodmap
works for this when run manually but doesn't work when run as an i3
startup app for some reason. Xmodmap seems to be deprecated anyway.

Doing this with xkb should be simpler, and may well be, but several
hours of research left me empty handed.

So this awful kludge is used.