[emacs] Load system and host dependent init files
Load system and host dependent files from emacs.d if they
exist. The files are in the following formats:

init-host-<name>.el
init-system-<name>.el

"/" in names are converted to "-".  There are probably more cases
that should be handled, but this gets the most common one that would
cause issues.