[i3] Map super and hyper to separate modifiers
Define a local "symbols" file that can override defines in the system
file.  This is done by adding a "local" term to the current xkb map
symbols section and recompiling the xkb mapping for the display while
providing the local xkb directory so that the overriding file is picked
up.