[emacs] Fix rtags find at point fns
This was going to gtags at times when it shouldn't. Currently the only
issue is if the symbol at point is not known by rtags (e.g. in a
comment), then gtags will still be run even though rtags finds the
symbol and shows the right buffer.

rtags-last-request-not-indexed doesn't seem to capture this case, so
there is some more work to be done to fix this edge case.