[emacs] remove unneeded config
The latest rtags is needed for this to work. Since it implements XDG
basedir specification it will pick up our ~/.config/rdmrc and it may
have created ~/.config/rtags before us, so just symlink the config
file there. Also, it will use the ~/.cache dir by default so no need
to create it.