[emacs] Suppress echo area while minibuffer active
This works under the assumption that all messages that occur when
minibuffer is active are stupid background things like ispell
dictionary loading and can be safely ignored and that messages which
occur when executing a command directly *are* important and should
show. This works as well when a command is executed after selection
from minibuffer aka M-x, as the minibuffer is exited before the
command runs.

One edge case is when recursive minibuffers are enabled and e.g. M-x
is used to run a command which shows something important in the echo
area (perhaps needing input?). In this case the echo area will be
suppressed, however I'm not sure there is a use case for this.