[emacs] Fix peep-dired
This was surprisingly hard.  Peep-dired has a keymap
which tries to map SPC and C-SPC. This caused an error because
SPC is a prefix (I think).  The googles were consulted
and a great many incantations were invoked, but what worked is
to simply redefine the offending keymap.

Also a struggle was to get j and k to work correctly.
Finally the nuclear option of evil-make-intercept was the correct
heavy tool to use.  This makes the given keymap override every
evil binding, and things now "just" work.

Also, moved peep-dired and dired-narrow out from dired use-package,
because they are key-binding invoked and can be demand loaded.