[emacs] Remove smart-tabs-mode
There are a number of issues that this mode triggers, and for
relatively little benefit. Arg alignment with spaces looks nice, but
creates larger diffs for refactoring than necessary. When needed
because of line length (but only then), it is better to indent all
arguments once or twice, this is readable and is diff friendly.