[emacs] Use evil-local-mode for hooks
Use evil-local-mode instead of evil-mode for the mode hooks.
Doing the latter can cause issues when someone uses a derived mode on
a background buffer. When it does that, evil-mode is called. This can
reinitialize evil and switch to normal state. This was seen with eglot
which uses a background gfm-mode buffer to font-lock markdown text.