[emacs] Workaround company argument expansion with tng mode
In tng mode any non-command key will call company-complete-selection
but suppress the backend which would normally fill in arguments. This
fixes that by mapping ( to company-complete-selection directly, since
for arguments in many languages, ( is going to be the next thing typed
anyway. Even in other languages, ( can still be used, the parameters
and surrounding syntax is added by the backend.

It's not clear to me why company does the suppression, but this works
around it.