[emacs] Terminal friendly keybindings
* Modify keybindings to work in terminal
* turn on mouse mode
* handle funky key mapping in terminal mode

Note on the funky key mapping: These are currently necessary to handle
xterm's mapping of certain characters. This is part of the work around
for the historic bad handling of certain keys by terminals, and
unfortunately xterm seems to be one of the few that convert keys in
the newer way. Emacs is also deficient in that it doesn't map *some*
of these sequences, but this code adds those mappings. This is all as
I understand it.