[emacs] No more jumpy cursor
These changes seem to tame the jumpy cursor syndrome. Sublimity
was nice, but I can't get it to work with the other changes, and
it wasn't _that_ nice.

Ironically, smooth-scrolling package or its config was causing a jump when cursor
was at edge of screen and cursor was moved toward edge.  Not sure what
it buys me so out it goes.