[emacs] Return C-u to evil
Some packages with vi bindings use only C-u/C-d for scrolling (instead
of C-f/C-b), and for consistency/muscle memory it's better to use
these in emacs. To do this, move C-u to M-u, and enable the C-u
behaviors in evil. M-u is upcase-word in emacs mode, which is not
missed.