[emacs] Fix projectile project finding again
Make sure projectile-project-root-files is set to one thing not
umpteen different marker file names. Otherwise it's impossible to use
projectile with projects that happen to have one of those marker files
some where in their tree. Once you edit a file in a directory with one
of the marker files, the top of the projectile tree is then set to
that directory.  This messes up other functions that rely on the
projectile root being the root of the project you are working on.

This commit doesn't entirely fix it.  If you start emacs in a
subdirectory with one of the markers, that directory becomes
the project root.  I haven't been able to fix this so far.
The workaround is to start emacs at the root of the project.