[emacs] Use project and popper
* Remove projectile, add project
* Remove popwin, add popper.
* Remove shell-pop
* Remove magit fullscreen status hack

NOTE Shackle was not added because it was messing
with (window-prev-buffers), which I used for rotating among MRU
buffers, and author has moved all his code to a host with no issue
reporting capability, so it's not going to get fixed.

For future reference, the bug can be reproduced by starting emacs and
using find-file to load a file. Evaluating (window-prev-buffers)
returns nil, but should have the buffer that is being viewed. This is
the result with emacs -Q and also when shackle is disabled.