[emacs] Super key on windows
This makes the windows key super, and this works for most
super bindings, s-l and s-g are the only exceptions
I know about currently.

For this to work on windows, use powertoys keyboard manager
to remap win (left) to disable, and caps lock to win.
For both of these use the Type button for the left hand
side, and the popup for the right hand side.  Ignore
warnings on OK.

Caps-lock will then become the windows key for all but window key
bindings where it is held down, which are not many.
Win-l and win-g are two I know about.