[emacs] Fix mis-ordered paste issue
If external program is used to copy text into the clipboard and then a
line is killed in emacs, the kill ring has the external text first and
then the killed line, although I'm not sure why the external text is
there at all. This change causes the order to be reversed, which is
more expected behavior, as it corresponds to the temporal order of the
events.