[emacs] Let emacs kills/evil yanks pollute the system clipboard
This can be annoying, but there doesn't seem to be a way to only
let emacs yank/evil put come from the system clipboard.

If interprogram-cut-function is set to nil, it almost works, but there
is an interaction with save-interprogram-paste-before-kill=t, where
the killed text is actually after the system clipboard text in the
kill ring, which is the opposite of desired. I believe this has to do
with some variables that the default interprogram-cut-function sets
that prevent this. It is probably solvable, but I ran out of interest.